Add cpp files & code refactor
