+++ physics/include/rawrbox/physics/utils.hpp
@@ -1,5 +1,4 @@
 #pragma once
-#include <rawrbox/math/pi.hpp>
 #include <rawrbox/math/vector3.hpp>
 #include <rawrbox/math/vector4.hpp>
 
@@ -8,12 +7,7 @@
 namespace rawrbox {
 	class PhysUtils {
 	public:
-		static rawrbox::Vector3f posToVec(const JPH::Vec3& q) {
-			return {q.GetX(), q.GetY(), q.GetZ()};
-		}
-
-		static rawrbox::Vector4f quatToVec4(const JPH::Quat& q) {
-			return {q.GetX(), q.GetY(), q.GetZ(), q.GetW()};
-		}
+		static rawrbox::Vector3f posToVec(const JPH::Vec3& q);
+		static rawrbox::Vector4f quatToVec4(const JPH::Quat& q);
 	};
 } // namespace rawrbox
