+++ render/include/rawrbox/render/postprocess/bloom.hpp
@@ -1,22 +1,7 @@
 #pragma once
 #include <rawrbox/render/postprocess/base.hpp>
-#include <rawrbox/render/shader_defines.hpp>
-#include <rawrbox/render/static.hpp>
-#include <rawrbox/render/util/uniforms.hpp>
-
-#include <generated/shaders/render/all.hpp>
 
 #include <bgfx/bgfx.h>
-#include <bx/math.h>
-
-#include <memory>
-
-// NOLINTBEGIN(*)
-static const bgfx::EmbeddedShader bloom_shaders[] = {
-    BGFX_EMBEDDED_SHADER(vs_post_bloom),
-    BGFX_EMBEDDED_SHADER(fs_post_bloom),
-    BGFX_EMBEDDED_SHADER_END()};
-// NOLINTEND(*)
 
 namespace rawrbox {
 	class PostProcessBloom : public rawrbox::PostProcessBase {
@@ -27,39 +12,17 @@ namespace rawrbox {
 		float _intensity = 0.08F;
 
 	public:
-		explicit PostProcessBloom(float intensity) : _intensity(intensity) {}
-
+		explicit PostProcessBloom(float intensity);
 		PostProcessBloom(PostProcessBloom&&) = delete;
 		PostProcessBloom& operator=(PostProcessBloom&&) = delete;
 		PostProcessBloom(const PostProcessBloom&) = delete;
 		PostProcessBloom& operator=(const PostProcessBloom&) = delete;
 
-		~PostProcessBloom() override {
-			RAWRBOX_DESTROY(this->_program);
-			RAWRBOX_DESTROY(this->_bloom_intensity);
-		}
+		~PostProcessBloom() override;
 
-		virtual void setIntensity(float in) {
-			this->_intensity = in;
-			if (bgfx::isValid(this->_bloom_intensity)) rawrbox::UniformUtils::setUniform(this->_bloom_intensity, this->_intensity);
-		}
+		virtual void setIntensity(float in);
 
-		void upload() override {
-			// Load Shader --------
-			bgfx::RendererType::Enum type = bgfx::getRendererType();
-			bgfx::ShaderHandle vsh = bgfx::createEmbeddedShader(bloom_shaders, type, "vs_post_bloom");
-			bgfx::ShaderHandle fsh = bgfx::createEmbeddedShader(bloom_shaders, type, "fs_post_bloom");
-
-			this->_program = bgfx::createProgram(vsh, fsh, true);
-			if (!bgfx::isValid(this->_program)) throw std::runtime_error("[RawrBox-Bloom] Failed to initialize shader program");
-			// ------------------
-
-			this->_bloom_intensity = bgfx::createUniform("u_intensity", bgfx::UniformType::Vec4, 1);
-			rawrbox::UniformUtils::setUniform(this->_bloom_intensity, this->_intensity);
-		}
-
-		void applyEffect() override {
-			bgfx::submit(rawrbox::CURRENT_VIEW_ID, this->_program);
-		}
+		void upload() override;
+		void applyEffect() override;
 	};
 } // namespace rawrbox
