+++ render/include/rawrbox/render/resources/assimp/model.hpp
@@ -7,32 +7,16 @@ namespace rawrbox {
 	class ResourceAssimp : public rawrbox::Resource {
 	public:
 		std::shared_ptr<rawrbox::AssimpImporter> model = nullptr;
-		~ResourceAssimp() override { this->model.reset(); }
 
-		bool load(const std::vector<uint8_t>& buffer) override {
-			this->model = std::make_shared<rawrbox::AssimpImporter>(flags);
-			this->model->load(this->filePath, buffer);
-
-			return true;
-		}
+		~ResourceAssimp() override;
+		bool load(const std::vector<uint8_t>& buffer) override;
 	};
 
 	class AssimpLoader : public rawrbox::Loader {
 	public:
 		~AssimpLoader() override = default;
 
-		std::unique_ptr<rawrbox::Resource> createEntry(uint32_t flags = 0) override {
-			return std::make_unique<rawrbox::ResourceAssimp>();
-		}
-
-		bool canLoad(const std::string& fileExtention) override {
-			return fileExtention == ".obj" ||
-			       fileExtention == ".fbx" ||
-			       fileExtention == ".dae" ||
-			       fileExtention == ".3ds" ||
-			       fileExtention == ".mdl" ||
-			       fileExtention == ".gltf" ||
-			       fileExtention == ".md5mesh";
-		}
+		std::unique_ptr<rawrbox::Resource> createEntry(uint32_t flags = 0) override;
+		bool canLoad(const std::string& fileExtention) override;
 	};
 } // namespace rawrbox
