+++ render/include/rawrbox/render/resources/font.hpp
@@ -3,33 +3,19 @@
 #include <rawrbox/render/text/engine.hpp>
 #include <rawrbox/resources/loader.hpp>
 
-#include <map>
-
 namespace rawrbox {
 	class ResourceFont : public rawrbox::Resource {
 
 	public:
-		std::shared_ptr<rawrbox::Font> getSize(int size, uint32_t flags = 0) {
-			return rawrbox::TextEngine::load(filePath.generic_string(), size);
-		}
+		std::shared_ptr<rawrbox::Font> getSize(int size, uint32_t flags = 0);
 	};
 
 	class FontLoader : public rawrbox::Loader {
 	public:
 		FontLoader() = default;
-		~FontLoader() override { rawrbox::TextEngine::shutdown(); };
+		~FontLoader() override;
 
-		std::unique_ptr<rawrbox::Resource> createEntry(uint32_t flags = 0) override {
-			return std::make_unique<rawrbox::ResourceFont>();
-		}
-
-		bool canLoad(const std::string& fileExtention) override {
-			return fileExtention == ".ttf" ||
-			       fileExtention == ".otf" ||
-			       fileExtention == ".ttc" ||
-			       fileExtention == ".pfa" ||
-			       fileExtention == ".pfb" ||
-			       fileExtention == ".ttc";
-		}
+		std::unique_ptr<rawrbox::Resource> createEntry(uint32_t flags = 0) override;
+		bool canLoad(const std::string& fileExtention) override;
 	};
 } // namespace rawrbox
