+++ render/include/rawrbox/render/resources/texture.hpp
@@ -7,32 +7,17 @@ namespace rawrbox {
 	class ResourceTexture : public rawrbox::Resource {
 	public:
 		std::shared_ptr<rawrbox::TextureImage> texture = nullptr;
-		~ResourceTexture() override { this->texture.reset(); }
+		~ResourceTexture() override;
 
-		bool load(const std::vector<uint8_t>& buffer) override {
-			this->texture = std::make_shared<rawrbox::TextureImage>(this->filePath, buffer);
-			return true;
-		}
-
-		void upload() override {
-			if (this->texture == nullptr) return;
-			this->texture->upload();
-		}
+		bool load(const std::vector<uint8_t>& buffer) override;
+		void upload() override;
 	};
 
 	class TextureLoader : public rawrbox::Loader {
 	public:
 		~TextureLoader() override = default;
 
-		std::unique_ptr<rawrbox::Resource> createEntry(uint32_t flags = 0) override {
-			return std::make_unique<rawrbox::ResourceTexture>();
-		}
-
-		bool canLoad(const std::string& fileExtention) override {
-			return fileExtention == ".png" ||
-			       fileExtention == ".bmp" ||
-			       fileExtention == ".jpg" ||
-			       fileExtention == ".jpeg";
-		}
+		std::unique_ptr<rawrbox::Resource> createEntry(uint32_t flags = 0) override;
+		bool canLoad(const std::string& fileExtention) override;
 	};
 } // namespace rawrbox
