+++ resources/include/rawrbox/resources/loaders/json.hpp
@@ -14,22 +14,14 @@ namespace rawrbox {
 
 	public:
 		std::shared_ptr<nlohmann::json> json = nullptr;
-		~ResourceJSON() override { this->json.reset(); }
+		~ResourceJSON() override;
 
-		bool load(const std::vector<uint8_t>& buffer) override {
-			json = std::make_shared<nlohmann::json>(nlohmann::json::parse(buffer));
-			return true;
-		}
+		bool load(const std::vector<uint8_t>& buffer) override;
 	};
 
 	class JSONLoader : public rawrbox::Loader {
 	public:
-		std::unique_ptr<rawrbox::Resource> createEntry(uint32_t flags = 0) override {
-			return std::make_unique<rawrbox::ResourceJSON>();
-		}
-
-		bool canLoad(const std::string& fileExtention) override {
-			return fileExtention == ".json";
-		}
+		std::unique_ptr<rawrbox::Resource> createEntry(uint32_t flags = 0) override;
+		bool canLoad(const std::string& fileExtention) override;
 	};
 } // namespace rawrbox
