+++ resources/src/resource.cpp
@@ -1,0 +1,7 @@
+
+#include <rawrbox/resources/resource.hpp>
+
+namespace rawrbox {
+	bool Resource::load(const std::vector<uint8_t>& buffer) { return true; };
+	void Resource::upload(){};
+} // namespace rawrbox
