+++ utils/include/rawrbox/utils/time.hpp
@@ -1,0 +1,10 @@
+#pragma once
+#include <cstdint>
+
+namespace rawrbox {
+	class TimeUtils {
+	public:
+		static int64_t time();
+		static int64_t curtime();
+	};
+} // namespace rawrbox
